unify average_run script
